+++ src/main/kotlin/hw8/IdCard.kt
@@ -1,0 +1,7 @@
+package hw8
+
+import java.util.*
+
+class IdCard(var number: Int) {
+    private var dateExpire: Date? = null
+}
