+++ src/main/kotlin/hw8/Man.kt
@@ -1,0 +1,6 @@
+package hw8
+
+abstract class Man {
+    protected var surname: String? = null
+    protected var name: String? = null
+}
