+++ src/main/kotlin/hw8/PastPosition.kt
@@ -1,0 +1,6 @@
+package hw8
+
+data class PastPosition(
+    val name: String,
+    val department: Department
+)
