+++ src/main/kotlin/hw8/Room.kt
@@ -1,0 +1,5 @@
+package hw8
+
+data class Room(
+    val number: Int
+)
