+++ src/main/kotlin/hw8/Unit.kt
@@ -1,0 +1,5 @@
+package hw8
+
+interface Unit {
+    var personCount: Int
+}
